requirements.txt
asgiref==3.3.4
Django==3.2.4
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
